# docs/swagger.yaml
definitions:
  domain.Product:
    properties:
      id:
        type: string
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
      sku:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /products:
    get:
      description: Возвращает список товаров с фильтрацией по цене и поддержкой страниц
      parameters:
      - description: Лимит (по умолчанию 10)
        in: query
        name: limit
        type: integer
      - description: Сдвиг (по умолчанию 0)
        in: query
        name: offset
        type: integer
      - description: Минимальная цена
        in: query
        name: minPrice
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Product'
            type: array
        "500":
          description: internal server error
          schema:
            type: string
      summary: Список товаров с пагинацией
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Добавляет новый товар в базу данных склада
      parameters:
      - description: Данные товара
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/domain.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Product'
        "400":
          description: invalid request body
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Создать новый товар
      tags:
      - products
  /products/{sku}:
    get:
      description: Возвращает полную информацию о товаре по его уникальному артикулу
      parameters:
      - description: Артикул товара (SKU)
        in: path
        name: sku
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Product'
        "404":
          description: product not found
          schema:
            type: string
      summary: Получить товар по SKU
      tags:
      - products
  /products/analytics:
    get:
      description: 'Возвращает агрегированную статистику: общее количество, стоимость
        и среднюю цену'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: internal server error
          schema:
            type: string
      summary: Аналитика склада
      tags:
      - analytics
  /products/transfer:
    get:
      description: Переводит количество товара с одного SKU на другой внутри атомарной
        транзакции
      parameters:
      - description: SKU отправителя
        in: query
        name: from
        required: true
        type: string
      - description: SKU получателя
        in: query
        name: to
        required: true
        type: string
      - description: Количество (по умолчанию 1)
        in: query
        name: amount
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Transaction success
          schema:
            type: string
        "500":
          description: error description
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Перевод остатков (Транзакция)
      tags:
      - inventory
swagger: "2.0"